requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
pydantic
tqdm
openai
openai-whisper
python-dotenv
google-generativeai

